Update the texture provider to current LibX
